fix #47 (#48)
This was a combination of me using Next.js canary and the nodejs
runtime, with betterauth.

Somehow the combination of that, and next-safe-client broke every single
server action, rendering the site basically useless.
This hopefully fixes that, and if not, point and shame me for the PR
that comes after this one.

---------

Signed-off-by: Omar Zunic <contact@omarzunic.com>